requirements.txt
appdirs==1.4.0
botocore==1.5.10
chalice==0.5.1
click==6.2
decorator==4.0.11
docutils==0.13.1
FormEncode==1.3.1
IMDbPY==5.1.1
jmespath==0.9.1
lxml==3.7.2
packaging==16.8
pbr==1.10.0
PyDispatcher==2.0.5
pyparsing==2.1.10
python-dateutil==2.6.0
six==1.10.0
SQLAlchemy==1.1.5
sqlalchemy-migrate==0.10.0
SQLObject==3.1.0
sqlparse==0.2.2
Tempita==0.5.2
typing==3.5.2.2
virtualenv==15.1.0
